Switch features to uint_8t

+++ mettagrid/mettagrid/grid_object.hpp
@@ -52,7 +52,7 @@ public:
     init(type_id, GridLocation(r, c, layer));
   }
 
-  virtual void obs(ObsType* obs, const vector<unsigned int>& offsets) const = 0;
+  virtual void obs(ObsType* obs, const vector<uint8_t>& offsets) const = 0;
 };
 
 #endif  // GRID_OBJECT_HPP
